requirements.txt
python-dotenv==1.0.0
gql==3.4.1
aiohttp==3.8.5
requests==2.31.0
openai-whisper==20231117
torch==2.2.0
numpy==1.26.4
pyannote-audio==3.1.1 